+++ src/generate/gen_calendar_ctrl.cpp
@@ -50,7 +50,12 @@ bool CalendarCtrlGenerator::SettingsCode(Code& code)
 {
     if (code.IsTrue(prop_focus))
     {
-        code.NodeName().Function("SetFocus(").EndFunction();
+        auto form = code.node()->getForm();
+        // wxDialog and wxFrame will set the focus to this control after all controls are created.
+        if (!form->isGen(gen_wxDialog) && !form->isGen(gen_wxFrame))
+        {
+            code.Eol(eol_if_empty).NodeName().Function("SetFocus(").EndFunction();
+        }
     }
     return true;
 }