+++ src/generate/styled_text.cpp
@@ -928,7 +928,12 @@ bool StyledTextGenerator::SettingsCode(Code& code)
 
     if (code.IsTrue(prop_focus))
     {
-        code.Eol(eol_if_needed).NodeName().Function("SetFocus(").EndFunction();
+        auto form = code.node()->getForm();
+        // wxDialog and wxFrame will set the focus to this control after all controls are created.
+        if (!form->isGen(gen_wxDialog) && !form->isGen(gen_wxFrame))
+        {
+            code.Eol(eol_if_needed).NodeName().Function("SetFocus(").EndFunction();
+        }
     }
 
     // REVIEW: [Randalphwa - 12-28-2022] The caller closes the brace -- but it makes more sense